feat: add android native logs
