+++ README.md
@@ -10,6 +10,29 @@ npx pod-install
 
 ## Usage
 
+BREAKING CHANGE
+Starting on version 0.5 we add the native logs also to the file. This means the configuration has to be done also on the native side.
+
+### iOS
+Not yet supported
+
+### Android
+```kotlin
+import com.mattermost.turbolog.TurboLog
+import com.mattermost.turbolog.ConfigureOptions
+
+class MainApplication : NavigationApplication(), INotificationsApplication {
+    // ...
+    override fun onCreate() {
+        // ...
+        TurboLog.configure(options = ConfigureOptions())
+        TurboLog.i("Label", "Text")
+    }
+}
+```
+Configure the application on application create. Use `TurboLog.X` instead of `Log.X` on the native side to get the logs show in the files.
+
+### JavaScript
 ```js
 import TurboLogger from "@mattermost/react-native-turbo-log";
 
@@ -27,17 +50,19 @@ You can configure TurboLogger to customize the rolling policy, the interception 
 
 #### TurboLogger.configure(options?: ConfigureOptions) => Promise<void>
 
+Starting on 0.5, configuration has to be made at native level, and extra configuration can be given at JavaScript level.
+
 Initialize the TurboLogger with the specified options. Once the configure promise is resolved, all `console` calls are captured and writen to the log file unless specified otherwise. To ensure that no logs are missing, it is good practice to `await` this call at the launch of your app.
 
-| Option | Description | Default |
-| --- | --- | --- |
-| `logLevel` | Minimum log level for file output (it won't affect console output) | LogLevel.Debug |
-| `captureConsole` | If `true`, all `console` calls are automatically captured and written to a log file  | `true` |
-| `dailyRolling` | If `true`, a new log file is created every day | `false` |
-| `maximumFileSize` | A new log file is created when current log file exceeds the given size in bytes. Set it to `0` to disable | `1024 * 1024` (1MB) |
-| `maximumNumberOfFiles` | Maximum number of log files to keep. When a new log file is created, if the total number of files exceeds this limit, the oldest file is deleted. `0` to disable | `5` |
-| `logsDirectory` | Absolute path of directory where log files are stored. If not defined, log files are stored in the cache directory of the app | `undefined` |
-| `logToFile` | If `true`, log files are created and written to the filesystem. It can also be changed by calling the `setLogToConsole` method | `true` |
+| Option | Description | Default | Set at |
+| --- | --- | --- | --- |
+| `logLevel` | Minimum log level for file output (it won't affect console output) | LogLevel.Debug | Native |
+| `captureConsole` | If `true`, all `console` calls are automatically captured and written to a log file  | `true` | JavaScript |
+| `dailyRolling` | If `true`, a new log file is created every day | `false` | Native |
+| `maximumFileSize` | A new log file is created when current log file exceeds the given size in bytes. Set it to `0` to disable | `1024 * 1024` (1MB) | Native |
+| `maximumNumberOfFiles` | Maximum number of log files to keep. When a new log file is created, if the total number of files exceeds this limit, the oldest file is deleted. `0` to disable | `5` | Native |
+| `logsDirectory` | Absolute path of directory where log files are stored. If not defined, log files are stored in the cache directory of the app | `undefined` | Native |
+| `logToFile` | If `true`, log files are created and written to the filesystem. It can also be changed by calling the `setLogToConsole` method | `true` | Native |
 
 #### TurboLogger.deleteLogs(): Promise<boolean>
